+++ lib/project_types/script/layers/domain/script_config.rb
@@ -4,17 +4,15 @@ module Script
   module Layers
     module Domain
       class ScriptConfig
-        attr_reader :content, :version, :title, :description, :configuration_ui, :configuration, :filename
+        attr_reader :content, :version, :configuration_ui, :configuration, :filename
 
-        REQUIRED_FIELDS = %w(version title)
+        REQUIRED_FIELDS = %w(version)
 
         def initialize(content:, filename:)
           @filename = filename
           validate_content!(content)
           @content = content
           @version = @content["version"].to_s
-          @title = @content["title"]
-          @description = @content["description"]
           @configuration_ui = @content.fetch("configurationUi", true)
           @configuration = @content["configuration"]
         end